+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Changed
 
 - 上映イベント予約の所有権検索時に使用するMongoDBインデックスを追加。
+- 'aws.cognito.signin.user.admin'スコープで会員インターフェースを利用可能にするよう対応。
 
 ### Deprecated